refactor: buildProfileManagerOptionsFromPartialPersistedProfile
