+++ packages/shared/lib/contexts/onboarding/actions/initialiseProfileManagerFromOnboardingProfile.ts
@@ -1,7 +1,7 @@
 import { get } from 'svelte/store'
 
 import {
-    buildProfileManagerOptionsFromProfileData,
+    buildProfileManagerOptionsFromPartialPersistedProfile,
     initialiseProfileManager,
     profileManager,
 } from '@core/profile-manager'
@@ -9,10 +9,7 @@ import {
 import { onboardingProfile, updateOnboardingProfile } from '../stores'
 import { removeProfileFolder } from '@core/profile'
 
-export async function initialiseProfileManagerFromOnboardingProfile(
-    checkForExistingManager?: boolean,
-    chainId?: number
-): Promise<void> {
+export async function initialiseProfileManagerFromOnboardingProfile(checkForExistingManager?: boolean): Promise<void> {
     const existingManager = get(profileManager)
     if (existingManager) {
         if (!checkForExistingManager) {
@@ -23,11 +20,7 @@ export async function initialiseProfileManagerFromOnboardingProfile(
         }
     }
 
-    const profileManagerOptions = await buildProfileManagerOptionsFromProfileData(
-        get(onboardingProfile),
-        get(onboardingProfile).network?.id,
-        chainId
-    )
+    const profileManagerOptions = await buildProfileManagerOptionsFromPartialPersistedProfile(get(onboardingProfile))
     const { storagePath, coinType, clientOptions, secretManager } = profileManagerOptions
     const { id } = get(onboardingProfile)
     const manager = await initialiseProfileManager(storagePath, coinType, clientOptions, secretManager, id)
