+++ packages/shared/lib/core/network/constants/index.ts
@@ -10,3 +10,4 @@ export * from './network-status-poll-interval.constant'
 export * from './network.constant'
 export * from './official-node-urls.constant'
 export * from './seconds-per-milestone.constant'
+export * from './testnet-coin-type.constant'
