+++ packages/shared/lib/core/network/constants/testnet-coin-type.constant.ts
@@ -1,0 +1,1 @@
+export const TESTNET_COIN_TYPE = 1
