+++ packages/shared/lib/core/profile-manager/utils/index.ts
@@ -1,4 +1,4 @@
-export * from './buildProfileManagerOptionsFromProfileData'
+export * from './buildProfileManagerOptionsFromPartialPersistedProfile'
 export * from './getSecretManagerFromProfileType'
 export * from './getSecretManagerPath'
 export * from './validateWalletApiEvent'
